% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.acs.employment-package.R
\docType{package}
\name{dc.acs.employment-package}
\alias{dc.acs.employment}
\alias{dc.acs.employment-package}
\title{dc.acs.employment: Social Data Commons: Education & Training: Employment Rate}
\description{
Allows user to easily get employment rate data from the SDAD Data Commons. Data can be provided directly or in a file.
}
\author{
\strong{Maintainer}: Hanna Charankevich \email{hc2cc@virginia.edu}

Authors:
\itemize{
  \item Social and Decision Analytics Division (SDAD), Biocomplexity Institute, University of Virginia
}

}
\keyword{internal}
